refactor: move sort-switcher-tasks from shared/ into project-switcher feature
