+++ src/renderer/features/project-switcher/sort-switcher-tasks.ts
